refactor: userIdとesIdの型をそれぞれstringとlongからintへ変更（確定）

+++ backend/src/main/java/chibainfo5/es_manager/domain/EntrysheetsKey.java
@@ -5,6 +5,6 @@ import lombok.Data;
 
 @Data
 public class EntrysheetsKey implements Serializable {
-    private Long userId;
-    private Long esId;
+    private int userId;
+    private int esId;
 }
